app/src/main/java/com/mentoriatiago/integramarketplace/usecases/GetAllSellers.java
package com.mentoriatiago.integramarketplace.usecases;
import com.mentoriatiago.integramarketplace.domains.Seller;
import com.mentoriatiago.integramarketplace.gateways.outputs.SellerDataGateway;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.stereotype.Component;

@Component
public class GetAllSellers {
    @Autowired
    private SellerDataGateway sellerDataGateway;

    public Page<Seller> getSellers(int pageNumber, int pageSize){
        PageRequest pageable = PageRequest.of(pageNumber, pageSize);
        return sellerDataGateway.findAll(pageable);
    }
}
